+++ modules/effects/arrow-effects-extensions/src/main/kotlin/arrow/effects/extensions/optiont.kt
@@ -1,11 +1,11 @@
 package arrow.effects.extensions
 
 import arrow.core.None
-import arrow.data.OptionT
-import arrow.data.OptionTOf
-import arrow.data.OptionTPartialOf
-import arrow.data.extensions.OptionTMonadError
-import arrow.data.value
+import arrow.mtl.OptionT
+import arrow.mtl.OptionTOf
+import arrow.mtl.OptionTPartialOf
+import arrow.mtl.extensions.OptionTMonadError
+import arrow.mtl.value
 import arrow.effects.Ref
 import arrow.effects.typeclasses.Async
 import arrow.effects.typeclasses.Bracket
