+++ modules/mtl/arrow-mtl-data/src/main/kotlin/arrow/mtl/OptionT.kt
@@ -1,4 +1,4 @@
-package arrow.data
+package arrow.mtl
 
 import arrow.Kind
 import arrow.core.Either
