+++ modules/ui/arrow-ui-data/build.gradle
@@ -11,7 +11,6 @@ dependencies {
     testRuntime("org.junit.vintage:junit-vintage-engine:$jUnitVintageVersion")
     testCompile "io.kotlintest:kotlintest-runner-junit4:$kotlinTestVersion"
     testCompile project(':arrow-test')
-    testCompile project(':arrow-syntax')
 }
 
 apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
